Successfully merged 5944 commits from upstream/master with the following strategy:

**Key Decisions:**
- Adopted upstream's modular BUILD.bazel architecture
- Updated to latest dependency versions (llvm 20.1.8, rules_cc 0.1.2, etc.)
- Added yaml-cpp and abseil-cpp from upstream
- Preserved bzlmodify-specific configurations (tcl, swig, llvm-project overrides)
- Kept WORKSPACE and MODULE.bazel.lock from upstream for compatibility
- Maintained .direnv support in ignore files

**Major Changes:**
- Merged MODULE.bazel: combined bzlmodify's custom overrides with upstream's latest dependencies
- Adopted upstream BUILD.bazel with internal module dependencies
- Resolved conflicts in source files by adopting upstream versions
- Updated bazel configuration files to match upstream structure

This merge maintains bzlmodify's goal of upstream-compatible bzlmod support while incorporating all upstream improvements.

+++ src/gpl/src/fft.cpp
@@ -7,14 +7,13 @@
 #include <cfloat>
 #include <cmath>
 #include <cstdlib>
-#include <iostream>
 #include <utility>
 
 #define REPLACE_FFT_PI 3.141592653589793238462L
 
 namespace gpl {
 
-FFT::FFT(int binCntX, int binCntY, int binSizeX, int binSizeY)
+FFT::FFT(int binCntX, int binCntY, float binSizeX, float binSizeY)
     : binCntX_(binCntX),
       binCntY_(binCntY),
       binSizeX_(binSizeX),
